refactor: bibleVersionSelect, cardHideOptionSelect 별도 feature로 분리
