+++ src/mock/handler.ts
@@ -8,7 +8,7 @@ import {
 import { SUPABASE_URL } from '@/lib/supabase/supabaseConfig';
 import { BIBLE_VERSIONS_LIST } from '@utils/constants';
 
-import { CARD_HIDE_OPTIONS_LIST } from '@features/drilling/constants/cardHideOptions';
+import { CARD_HIDE_OPTIONS_LIST } from '@features/cardHideOptionSelect/constants/cardHideOptions';
 import { SORT_METHODS_LIST } from '@features/exam/constants/sortMethods';
 
 const baseURL = `${SUPABASE_URL}/rest/v1`;
